requirements.txt
audonnx==0.6.2
cn2an==0.5.19
eng_to_ipa==0.0.2
indic_transliteration==2.3.40
inflect==6.0.2
jamo==0.4.1
ko_pron==1.3
librosa==0.9.2
num_thai==0.0.5
numba==0.56.4
numpy==1.23.5
OpenCC==1.1.1
openjtalk==0.3.0.dev2
pandas==1.5.2
Pillow==9.3.0
pyopenjtalk==0.3.0
pypinyin==0.47.1
requests==2.28.1
scipy==1.9.3
torch==1.12.1+cu116
Unidecode==1.3.6
